adding limitation test and adding linitation conditional

+++ app/controllers/api/v1/listings_controller.rb
@@ -18,14 +18,19 @@ class Api::V1::ListingsController < ApplicationController
 
   def create
     listing = current_user.listings.create(listing_params)
-    if listing.persisted? && attach_image(listing)
-      render json: { message: 'The listing has been created successfully!' }
-    elsif !attach_image(listing)
-      listing.destroy
-      render json: { message: "The image can't be blank"}, status:422
-    else 
-      render_error_message(listing.errors)
+    if current_user.listings.size < 3
+      if listing.persisted? && attach_image(listing)
+        render json: { message: 'The listing has been created successfully!' }
+      elsif !attach_image(listing)
+        listing.destroy
+        render json: { message: "The image can't be blank"}, status:422
+      else 
+        render_error_message(listing.errors)
+      end
+    else
+      render json: { message: "Subscribe to create more listing."}, status:422
     end
+
   rescue StandardError => e
     render json: { message: e }, status: 422
   end
